requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
anthropic>=0.8.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=7.4.0
tabulate>=0.9.0
